fix: Save security reports for static pods
Note that kublet can manage pods independently by reading pod definition
files from a configured host directory (typically /etc/kubernetes/manifests).
Such pods are called *static pods* and are owned by a cluster Node.
So far we could not save vulnerability and config audit reports for such pods.

In this commit we treat them as unmanaged pods. (Otherwise we'd require
cluster-scoped permissions to get Nodes in order to set the owner reference
when we create an instance of custom resource.)

Resolves: #234

Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>