refactor(conftest): Skip rescan when plugin ConfigMap is deleted (#489)
This wasn't an issue and we were handling ConfigMap deletion as
expected. However, I took this oportunity to refactor the code
a bit and use predicate instead of conditionally ingoring ConfigMap
events in the reconciliation loop. With predicate the events that
we're not interested in are not even added to the worker queue.

Resolves: #482