Feature only scan current revision of deployment
Add env vuln SCANNER_SCAN_ONLY_CURRENT_REVISIONS
If a replicaset is owned by a deployment we will compare the revision of the deployment
and replicaset to only generated vulnerbility reports for the current revision of the replicaset.

We need to check the owner of the replicaset and thanks to that we need access to get,watch,list a deployment.

Signed-off-by: Edvin Norling <edvin.norling@xenit.se>
Co-authored-by: Zach Stone <z.stone91@gmail.com>
Co-authored-by: Daniel Pacak <pacak.daniel@gmail.com>