feat: allow scanning only current revision of deployment (#870)
Adds the OPERATOR_VULNERABILITY_SCANNER_SCAN_ONLY_CURRENT_REVISIONS
environment variable to allow scanning only the current revision of
a Deployment. If a ReplicaSet is owned by a Deployment the code will check if
it's the current revision and scan it for vulnerabilities. Inactive ReplicaSets
will be ignored.

To check the owner of a ReplicaSet additional permissions to get, list, and
watch Deployments is required.

Resolves: #858
Resolves: #668

Signed-off-by: Edvin Norling <edvin.norling@xenit.se>
Co-authored-by: Zach Stone <z.stone91@gmail.com>
Co-authored-by: Daniel Pacak <pacak.daniel@gmail.com>