fix(helm): grant permissions to get,list,watch ciskubebenchreports (#1013)
Initially, we wanted to implement the least privileged principle
by creating Role or ClusterRole depending on install mode.
For example, we created Role if the target namespace was the same
as the operator namespace. However, this logic was too simple
because we have cluster-scoped CRDs, e.g. CISKubeBenchReport, that
require ClusterRole to grant RBAC permissions.

Resolves: #1001

Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>