feat(trivy): add mirror config in helm chart (#1159)
This patch adds the trivy mirror configuration to the starboard operator
helm chart. This particular implementation is made to match the
definitions from the documentation and therefore uses two layers for the
object and singular terminology.

The patch is required because currently there is no way than post
deployment patching of the configmap in order to utilise this feature,
which works against the intention of helm. It was also very confusing to
find the documentation providing this option, but the helm chart not
doing so, therefore this patch should help to cleanup potential
confusion.