feat: Support scan latest revision for replicationController from DeploymentConfig (#1352)
In OCP, there is a kind called DeploymentConfig, which manages the replicationController, and as we have revisionHistory for Deployments, similarly there are revision history for Deployment Config as well. Due to which many replication contollers will be present in the cluster. We dont want to scan them as they are not active resource.
So as part of this change, I have added following change
- consider "OPERATOR_CONFIG_AUDIT_SCANNER_SCAN_ONLY_CURRENT_REVISIONS" env var to limit scanning of old replication controller
- ingore pods created from deploymentConfig

resolves: 1351