opa version upgrade to latest (#1367)
To maintain compatibility with previous KE versions, we initially introduced the result. New feature to all Rego scripts on the server level.
However, we encountered an error within StarBoard. This problem arises from the utilization of "dots in rule heads," which is not compatible with the current OPA version employed by StarBoard.
As a resolution, updated the OPA version used within StarBoard to v0.46.1.
Reference: https://github.com/open-policy-agent/opa/issues/4653

Co-authored-by: Kiran Bodipi <kiran.bodipi@aquasec.com>